test(Bonus) : check unvalidated range
